Add handling for git repos in extra-deps
This complicates things somewhat and made it necessary to make
`StackConfig` an opaque type with accessors to get the separate
components in a more logical manner. For instance the `stackGitRepos`
accessor returns all git repos, regardless of whether they were
listed in the stack.yaml file in the 'extra-deps' section of the
'locations' section. Similarly, `stackExtraDeps` only returns
package names with versions.