Jenga.Git: Improve the way git is invoked
Previously submodules were updated and specific git commits checked out
by changing into the submodule and issuing the command. This requires
two system calls from Haskell code, the first to change into the directory
and the second to change back out.

The above scheme was dropped in favour of issing the commands in place
as:

    git -C <target-directory> <command>